+++ src/config/template.rs
@@ -6,11 +6,7 @@ pub fn generate_default_config() {
 }
 
 fn config_template() -> String {
-    let names_length = db::build_db()
-        .keys()
-        .map(|a| a.len())
-        .collect::<Vec<usize>>();
-    let max_name_length = names_length.iter().max().unwrap_or(&0) + 1;
+    let max_name_length: usize = db::build_db().keys().map(|a| a.len()).max().unwrap() + 1;
 
     let tools = db::fmt_tool_names_info(|(name, info)| {
         format!(