+++ src/install.rs
@@ -4,7 +4,7 @@ use crate::config::schema::Config;
 use crate::config::toml;
 use crate::infra::err;
 use crate::sync;
-use crate::sync::db::{fmt_tool_names, lookup_tool};
+use crate::sync::db::{build_db, fmt_tool_names_info, lookup_tool};
 
 /// Install a single tool
 pub fn install(config_path: PathBuf, name: String) {
@@ -16,7 +16,17 @@ fn install_tool(config: Config, name: String) {
     if let Some(tool_info) = lookup_tool(&name) {
         sync::sync_single_tool(config, name, tool_info.into());
     } else {
-        let tools = fmt_tool_names(|name| format!("    * {name}"));
+        let max_name_length: usize = build_db().keys().map(|a| a.len()).max().unwrap() + 1;
+
+        let tools = fmt_tool_names_info(|(name, info)| {
+            format!(
+                "    * {name} {delim:>padding$} https://github.com/{owner}/{repo}",
+                delim = "#",
+                padding = max_name_length - name.len(),
+                owner = info.owner,
+                repo = info.repo,
+            )
+        });
 
         let exit_message = format!(
             r#"Unknown tool: '{name}'