+++ src/sync/db.rs
@@ -118,21 +118,6 @@ pub fn build_db() -> BTreeMap<String, ToolInfo> {
     )
 }
 
-/// Format tool names of the database using a mutating formatting function
-/// The result is something like this (depending on a function)
-///
-/// ```toml
-/// # [bat]
-/// # [exa]
-/// ```
-pub fn fmt_tool_names<F: FnMut(&String) -> String>(fmt_tool: F) -> String {
-    build_db()
-        .keys()
-        .map(fmt_tool)
-        .collect::<Vec<String>>()
-        .join("\n")
-}
-
 /// Format tool names and info of the database using a mutating formatting function
 /// The result is something like this (depending on a function)
 ///